bkg/requirements.txt
pillow==10.2.0
pytest==7.4.4